Modify : checkValid 함수 utils 폴더로 분리
Modify : 상수 함수 바깥으로

Modify : 불필요한 버튼 삭제

Modify : 연도 입력 input width 수정
