+++ src/pages/ProjectStart/Components/CardHandleBtn.js
@@ -5,12 +5,9 @@ import styled from 'styled-components/macro';
 export default function CardHandleBtn({ toggleEditor }) {
   return (
     <ButtonWrapper>
-      <Button color="tertiaryGray" onClick={toggleEditor}>
-        <Icon className="fas fa-times" /> 닫기
+      <Button color="red" onClick={toggleEditor}>
+        <Icon className="fas fa-times" /> 저장 후 닫기
       </Button>
-      <CustomButton color="red">
-        <Icon className="fas fa-check" /> 저장하기
-      </CustomButton>
     </ButtonWrapper>
   );
 }
@@ -22,8 +19,4 @@ const ButtonWrapper = styled.div`
 
 const Icon = styled.i`
   margin-right: 5px;
-`;
-
-const CustomButton = styled(Button)`
-  margin-left: 10px;
 `;
